in preparation for Ravenfall v0.9, keeping everything up to date.

+++ src/RavenNest.DataModels/CharacterStatusEffect.cs
@@ -9,13 +9,13 @@ namespace RavenNest.DataModels
     {
         private Guid characterId;
         private StatusEffectType type;
-        private float amount;
+        private double amount;
         private DateTime startUtc;
         private DateTime expiresUtc;
 
         public Guid CharacterId { get => characterId; set => Set(ref characterId, value); }
         public StatusEffectType Type { get => type; set => Set(ref type, value); }
-        public float Amount { get => amount; set => Set(ref amount, value); }
+        public double Amount { get => amount; set => Set(ref amount, value); }
         public DateTime StartUtc { get => startUtc; set => Set(ref startUtc, value); }
         public DateTime ExpiresUtc { get => expiresUtc; set => Set(ref expiresUtc, value); }
     }
